@@ -1,9 +1,8 @@
 package com.fssa.learnJava.fop;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import org.junit.Test;
-
+import org.junit.jupiter.api.Test;
 /**
  * Unit test for simple App.
  */
