@@ -1,6 +1,6 @@
 package com.fssa.learnJava.fop.day04;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Assertions;
 
 public class TestCalculator {
